requirements.txt
aniso8601==10.0.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.10
decorator==4.4.2
Faker==30.4.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
google-auth==2.34.0
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
greenlet==3.1.1
idna==3.10
imageio==2.31.1
imageio-ffmpeg==0.4.9
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
moviepy==1.0.3
numpy==1.24.3
openpyxl==3.1.2
packaging==24.2
pandas==2.0.3
pika==1.3.2
Pillow==9.5.0
proglog==0.1.10
psycopg2-binary==2.9.9
pytest==8.3.4
pytest-mock==3.14.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.34
tqdm==4.66.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
gunicorn==21.2.0